set up the interval tabs component to trigger api call that fetch data specific to the interval
